fix: fixed new bill submission
Found multiple bugs with the filed handling. Fixed bill creation each time we change the file, now it is created only when we submit the form data. New bill is created with the file, and with a async/await combination we can retrieve fileUrl on creation then update the created bill with other form data